DEV Move DetectorBase as ABC in s/o/base.py
The DetectorBase class is now an AbstractBaseClass found
in serpentTools/objects/base.py
