+++ serpentTools/samplers/detector.py
@@ -11,7 +11,7 @@ from serpentTools.messages import (MismatchedContainersError, warning,
                                    SamplerError, SerpentToolsException)
 from serpentTools.plot import magicPlotDocDecorator
 from serpentTools.parsers.detector import DetectorReader
-from serpentTools.objects.containers import DetectorBase
+from serpentTools.objects.base import DetectorBase
 from serpentTools.samplers.base import Sampler, SampledContainer, SPREAD_PLOT_KWARGS
 
 class DetectorSampler(Sampler):
